requirements.txt
alembic==1.0.7
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
filelock==3.0.10
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.1.1
pytest-cov==2.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
toml==0.10.0
tox==3.7.0
urllib3==1.24.1
virtualenv==16.3.0
Werkzeug==0.14.1
